+++ codegen/end_to_end_test/lib/graphql/serializers.gql.g.dart
@@ -17,6 +17,8 @@ Serializers _$serializers = (new Serializers().toBuilder()
       ..add(GCreateReviewData.serializer)
       ..add(GCreateReviewData_createReview.serializer)
       ..add(GCreateReviewVars.serializer)
+      ..add(GDroidFragmentData.serializer)
+      ..add(GDroidFragmentVars.serializer)
       ..add(GEpisode.serializer)
       ..add(GHeroForEpisode.serializer)
       ..add(GHeroForEpisodeData.serializer)
@@ -46,7 +48,13 @@ Serializers _$serializers = (new Serializers().toBuilder()
       ..add(GReviewWithDateData.serializer)
       ..add(GReviewWithDateData_createReview.serializer)
       ..add(GReviewWithDateVars.serializer)
+      ..add(GcomparisonFieldsData.serializer)
+      ..add(GcomparisonFieldsData_friendsConnection.serializer)
+      ..add(GcomparisonFieldsData_friendsConnection_edges.serializer)
+      ..add(GcomparisonFieldsData_friendsConnection_edges_node.serializer)
       ..add(GcomparisonFieldsVars.serializer)
+      ..add(GheroDataData.serializer)
+      ..add(GheroDataVars.serializer)
       ..addBuilderFactory(
           const FullType(BuiltList, const [const FullType(DateTime)]),
           () => new ListBuilder<DateTime>())
@@ -64,7 +72,13 @@ Serializers _$serializers = (new Serializers().toBuilder()
             const FullType(GHeroWithFragmentsData_hero_friendsConnection_edges)
           ]),
           () => new ListBuilder<
-              GHeroWithFragmentsData_hero_friendsConnection_edges>()))
+              GHeroWithFragmentsData_hero_friendsConnection_edges>())
+      ..addBuilderFactory(
+          const FullType(BuiltList, const [
+            const FullType(GcomparisonFieldsData_friendsConnection_edges)
+          ]),
+          () =>
+              new ListBuilder<GcomparisonFieldsData_friendsConnection_edges>()))
     .build();
 
 // ignore_for_file: always_put_control_body_on_new_line,always_specify_types,annotate_overrides,avoid_annotating_with_dynamic,avoid_as,avoid_catches_without_on_clauses,avoid_returning_this,lines_longer_than_80_chars,omit_local_variable_types,prefer_expression_function_bodies,sort_constructors_first,test_types_in_equals,unnecessary_const,unnecessary_new