+++ examples/gql_example_build/lib/kitchen_sink/query.data.gql.g.dart
@@ -115,16 +115,16 @@ class _$GQueryOperationData_fieldSerializer
         ..add(serializers.serialize(object.id,
             specifiedType: const FullType(String)));
     }
-    if (object.bool$ != null) {
+    if (object.Gbool != null) {
       result
         ..add('bool')
-        ..add(serializers.serialize(object.bool$,
+        ..add(serializers.serialize(object.Gbool,
             specifiedType: const FullType(bool)));
     }
-    if (object.int$ != null) {
+    if (object.Gint != null) {
       result
         ..add('int')
-        ..add(serializers.serialize(object.int$,
+        ..add(serializers.serialize(object.Gint,
             specifiedType: const FullType(int)));
     }
     if (object.float != null) {
@@ -145,10 +145,10 @@ class _$GQueryOperationData_fieldSerializer
         ..add(serializers.serialize(object.scalar,
             specifiedType: const FullType(_i2.GScalar)));
     }
-    if (object.enum$ != null) {
+    if (object.Genum != null) {
       result
         ..add('enum')
-        ..add(serializers.serialize(object.enum$,
+        ..add(serializers.serialize(object.Genum,
             specifiedType: const FullType(_i2.GEnum)));
     }
     if (object.field != null) {
@@ -181,11 +181,11 @@ class _$GQueryOperationData_fieldSerializer
               specifiedType: const FullType(String)) as String;
           break;
         case 'bool':
-          result.bool$ = serializers.deserialize(value,
+          result.Gbool = serializers.deserialize(value,
               specifiedType: const FullType(bool)) as bool;
           break;
         case 'int':
-          result.int$ = serializers.deserialize(value,
+          result.Gint = serializers.deserialize(value,
               specifiedType: const FullType(int)) as int;
           break;
         case 'float':
@@ -201,7 +201,7 @@ class _$GQueryOperationData_fieldSerializer
               specifiedType: const FullType(_i2.GScalar)) as _i2.GScalar);
           break;
         case 'enum':
-          result.enum$ = serializers.deserialize(value,
+          result.Genum = serializers.deserialize(value,
               specifiedType: const FullType(_i2.GEnum)) as _i2.GEnum;
           break;
         case 'field':
@@ -458,9 +458,9 @@ class _$GQueryOperationData_field extends GQueryOperationData_field {
   @override
   final String id;
   @override
-  final bool bool$;
+  final bool Gbool;
   @override
-  final int int$;
+  final int Gint;
   @override
   final double float;
   @override
@@ -468,7 +468,7 @@ class _$GQueryOperationData_field extends GQueryOperationData_field {
   @override
   final _i2.GScalar scalar;
   @override
-  final _i2.GEnum enum$;
+  final _i2.GEnum Genum;
   @override
   final GQueryOperationData_field_field field;
   @override
@@ -495,12 +495,12 @@ class _$GQueryOperationData_field extends GQueryOperationData_field {
   _$GQueryOperationData_field._(
       {this.G__typename,
       this.id,
-      this.bool$,
-      this.int$,
+      this.Gbool,
+      this.Gint,
       this.float,
       this.string,
       this.scalar,
-      this.enum$,
+      this.Genum,
       this.field,
       this.idRequired,
       this.boolRequired,
@@ -564,12 +564,12 @@ class _$GQueryOperationData_field extends GQueryOperationData_field {
     return other is GQueryOperationData_field &&
         G__typename == other.G__typename &&
         id == other.id &&
-        bool$ == other.bool$ &&
-        int$ == other.int$ &&
+        Gbool == other.Gbool &&
+        Gint == other.Gint &&
         float == other.float &&
         string == other.string &&
         scalar == other.scalar &&
-        enum$ == other.enum$ &&
+        Genum == other.Genum &&
         field == other.field &&
         idRequired == other.idRequired &&
         boolRequired == other.boolRequired &&
@@ -604,12 +604,12 @@ class _$GQueryOperationData_field extends GQueryOperationData_field {
                                                                         G__typename
                                                                             .hashCode),
                                                                     id.hashCode),
-                                                                bool$.hashCode),
-                                                            int$.hashCode),
+                                                                Gbool.hashCode),
+                                                            Gint.hashCode),
                                                         float.hashCode),
                                                     string.hashCode),
                                                 scalar.hashCode),
-                                            enum$.hashCode),
+                                            Genum.hashCode),
                                         field.hashCode),
                                     idRequired.hashCode),
                                 boolRequired.hashCode),
@@ -626,12 +626,12 @@ class _$GQueryOperationData_field extends GQueryOperationData_field {
     return (newBuiltValueToStringHelper('GQueryOperationData_field')
           ..add('G__typename', G__typename)
           ..add('id', id)
-          ..add('bool\$', bool$)
-          ..add('int\$', int$)
+          ..add('Gbool', Gbool)
+          ..add('Gint', Gint)
           ..add('float', float)
           ..add('string', string)
           ..add('scalar', scalar)
-          ..add('enum\$', enum$)
+          ..add('Genum', Genum)
           ..add('field', field)
           ..add('idRequired', idRequired)
           ..add('boolRequired', boolRequired)
@@ -658,13 +658,13 @@ class GQueryOperationData_fieldBuilder
   String get id => _$this._id;
   set id(String id) => _$this._id = id;
 
-  bool _bool$;
-  bool get bool$ => _$this._bool$;
-  set bool$(bool bool$) => _$this._bool$ = bool$;
+  bool _Gbool;
+  bool get Gbool => _$this._Gbool;
+  set Gbool(bool Gbool) => _$this._Gbool = Gbool;
 
-  int _int$;
-  int get int$ => _$this._int$;
-  set int$(int int$) => _$this._int$ = int$;
+  int _Gint;
+  int get Gint => _$this._Gint;
+  set Gint(int Gint) => _$this._Gint = Gint;
 
   double _float;
   double get float => _$this._float;
@@ -678,9 +678,9 @@ class GQueryOperationData_fieldBuilder
   _i2.GScalarBuilder get scalar => _$this._scalar ??= new _i2.GScalarBuilder();
   set scalar(_i2.GScalarBuilder scalar) => _$this._scalar = scalar;
 
-  _i2.GEnum _enum$;
-  _i2.GEnum get enum$ => _$this._enum$;
-  set enum$(_i2.GEnum enum$) => _$this._enum$ = enum$;
+  _i2.GEnum _Genum;
+  _i2.GEnum get Genum => _$this._Genum;
+  set Genum(_i2.GEnum Genum) => _$this._Genum = Genum;
 
   GQueryOperationData_field_fieldBuilder _field;
   GQueryOperationData_field_fieldBuilder get field =>
@@ -737,12 +737,12 @@ class GQueryOperationData_fieldBuilder
     if (_$v != null) {
       _G__typename = _$v.G__typename;
       _id = _$v.id;
-      _bool$ = _$v.bool$;
-      _int$ = _$v.int$;
+      _Gbool = _$v.Gbool;
+      _Gint = _$v.Gint;
       _float = _$v.float;
       _string = _$v.string;
       _scalar = _$v.scalar?.toBuilder();
-      _enum$ = _$v.enum$;
+      _Genum = _$v.Genum;
       _field = _$v.field?.toBuilder();
       _idRequired = _$v.idRequired;
       _boolRequired = _$v.boolRequired;
@@ -778,12 +778,12 @@ class GQueryOperationData_fieldBuilder
           new _$GQueryOperationData_field._(
               G__typename: G__typename,
               id: id,
-              bool$: bool$,
-              int$: int$,
+              Gbool: Gbool,
+              Gint: Gint,
               float: float,
               string: string,
               scalar: _scalar?.build(),
-              enum$: enum$,
+              Genum: Genum,
               field: _field?.build(),
               idRequired: idRequired,
               boolRequired: boolRequired,