+++ codegen/gql_build/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.1
 - build var classes for fragments
+- build instantiable data classes for fragments
 
 ## 0.1.0