+++ codegen/gql_code_builder/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.1.1
 - build var classes for fragments
+- build instantiable data classes for fragments
+- consolidate serialization methods
 
 ## 0.1.0